sh: routes: renamed 'utils' in 'components'
